requirements.txt
tensorflow-gpu~=2.5.0
kaggle~=1.5.12
numpy~=1.19.5
pandas~=1.2.4
matplotlib~=3.4.2
APScheduler~=3.8.1
Flask~=1.1.4
marshmallow~=3.11.1
scikit-learn~=1.0.1
Pillow~=8.2.0
requests~=2.25.1
Flask-SQLAlchemy~=2.5.1
Werkzeug~=1.0.1
SQLAlchemy~=1.4.15
readerwriterlock~=1.0.9